+++ src/main/java/nl/dtls/fairdatapoint/Application.java
@@ -24,6 +24,7 @@ package nl.dtls.fairdatapoint;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
@@ -32,6 +33,7 @@ import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 @EnableWebMvc
 @EnableAsync
 @ComponentScan(basePackages = "nl.dtls.fairdatapoint.*")
+@ConfigurationPropertiesScan("nl.dtls.fairdatapoint.config.*")
 public class Application {
 
     public static void main(String[] args) {