+++ src/main/java/nl/dtls/fairdatapoint/api/dto/settings/SettingsUpdateDTO.java
@@ -20,24 +20,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.fairdatapoint.api.dto.reset;
+package nl.dtls.fairdatapoint.api.dto.settings;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import nl.dtls.fairdatapoint.entity.settings.SettingsMetricsEntry;
+
+import javax.validation.constraints.NotNull;
+import java.util.List;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
-public class ResetDTO {
+public class SettingsUpdateDTO {
 
-    private boolean users;
+    @NotNull
+    private List<SettingsMetricsEntry> metadataMetrics;
 
-    private boolean metadata;
-
-    private boolean resourceDefinitions;
-
-    private boolean settings;
+    @NotNull
+    private SettingsPingUpdateDTO ping;
 }