+++ src/main/java/nl/dtls/fairdatapoint/config/CacheConfig.java
@@ -45,6 +45,8 @@ public class CacheConfig {
 
     public static final String LABEL_CACHE = "LABEL_CACHE";
 
+    public static final String SETTINGS_CACHE = "SETTINGS_CACHE";
+
     @Bean
     public ConcurrentMapCacheManager cacheManager() {
         ConcurrentMapCacheManager cacheManager = new ConcurrentMapCacheManager();
@@ -54,6 +56,7 @@ public class CacheConfig {
                 RESOURCE_DEFINITION_CACHE,
                 RESOURCE_DEFINITION_PARENT_CACHE,
                 RESOURCE_DEFINITION_TARGET_CLASSES_CACHE,
+                SETTINGS_CACHE,
                 LABEL_CACHE
         ));
         return cacheManager;