+++ src/main/java/nl/dtls/fairdatapoint/service/reset/ResetService.java
@@ -27,9 +27,11 @@ import lombok.extern.slf4j.Slf4j;
 import nl.dtls.fairdatapoint.api.dto.reset.ResetDTO;
 import nl.dtls.fairdatapoint.database.mongo.repository.*;
 import nl.dtls.fairdatapoint.entity.resource.ResourceDefinition;
+import nl.dtls.fairdatapoint.entity.settings.Settings;
 import nl.dtls.fairdatapoint.service.metadata.exception.MetadataServiceException;
 import nl.dtls.fairdatapoint.service.metadata.generic.GenericMetadataService;
 import nl.dtls.fairdatapoint.service.resource.ResourceDefinitionCache;
+import nl.dtls.fairdatapoint.service.settings.SettingsService;
 import nl.dtls.fairdatapoint.vocabulary.DATACITE;
 import nl.dtls.fairdatapoint.vocabulary.FDP;
 import nl.dtls.fairdatapoint.vocabulary.R3D;
@@ -82,10 +84,6 @@ public class ResetService {
     private IRI language;
 
     @Autowired
-    @Qualifier("metadataMetrics")
-    private Map<String, String> metadataMetrics;
-
-    @Autowired
     protected Repository repository;
 
     @Autowired
@@ -119,11 +117,17 @@ public class ResetService {
     private GenericMetadataService genericMetadataService;
 
     @Autowired
+    private SettingsService settingsService;
+
+    @Autowired
     private MongoTemplate mongoTemplate;
 
     @PreAuthorize("hasRole('ADMIN')")
     public void resetToFactoryDefaults(ResetDTO reqDto) throws Exception {
         log.info("Resetting to factory defaults");
+        if (reqDto.isSettings()) {
+            settingsService.resetSettings();
+        }
         if (reqDto.isUsers() || reqDto.isMetadata()) {
             clearMemberships();
             restoreDefaultMemberships();
@@ -202,7 +206,6 @@ public class ResetService {
         try (RepositoryConnection conn = repository.getConnection()) {
             List<Statement> s = FactoryDefaults.repositoryStatements(
                     persistentUrl,
-                    metadataMetrics,
                     license,
                     language,
                     accessRightsDescription